test/sjis.dat

#data
<plaintext>
‚Ά‚λ‚Ν‚Ι‚Ω‚Φ‚Ζ
‚Ώ‚θ‚Κ‚ι‚π‚ν‚©
‚ζ‚½‚κ‚»‚Β‚Λ‚Θ
‚η‚ή‚¤‚ξ‚Μ‚¨‚­
‚β‚ά‚―‚Σ‚±‚¦‚Δ
‚ ‚³‚­‚δ‚ί‚έ‚µ
‚ο‚Π‚ΰ‚Ή‚·
#errors
#document
| <html>
|   <head>
|   <body>
|     <plaintext>
|       "
γ„γ‚γ―γ«γ»γΈγ¨
γ΅γ‚γ¬γ‚‹γ‚’γ‚γ‹
γ‚γγ‚γγ¤γ­γ
γ‚‰γ‚€γ†γ‚γ®γγ
γ‚„γΎγ‘γµγ“γγ¦
γ‚γ•γγ‚†γ‚γΏγ—
γ‚‘γ²γ‚‚γ›γ™"
